% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admiralci-package.R
\docType{package}
\name{admiralci-package}
\alias{admiralci}
\alias{admiralci-package}
\title{admiralci: ADaM in R Asset Library - CI/CD Workflows}
\description{
A set of common CI/CD workflows originally designed and developed for
    Admiral R package hosted in the Pharmaverse GitHub organization.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pharmaverse.github.io/admiralci}
  \item \url{https://github.com/pharmaverse/admiralci}
}

}
\author{
\strong{Maintainer}: Open Source \email{admiralci@pharmaverse.com}

}
\keyword{internal}
